% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_studyid}
\alias{make_studyid}
\title{Make study name}
\usage{
make_studyid(bridge, author, year)
}
\arguments{
\item{bridge}{Character vector of bridge id's.}

\item{author}{Character vector of authors.}

\item{year}{Character vector of years.}
}
\value{
Character vector of unique study id's.
}
\description{
Given bridge ID, author and year amek a unique study name. If author and year
exist use that adding a number to make unique (if duplicate author_year for)
a bridge ID. If author and year absent use bridge ID
}
