% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_unique}
\alias{make_unique}
\title{Add suffix to a character vector if it's longer than 1}
\usage{
make_unique(x, pad = "", suffix = letters)
}
\arguments{
\item{x}{Character vector to make unique}

\item{pad}{Characters to pad between string and suffix}

\item{suffix}{Vector of things to add, defalults to lower case letters}
}
\value{
A unique character vector
}
\description{
Add suffix to a character vector if it's longer than 1
}
