% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signif_base}
\alias{plot_signif_base}
\title{Plot factor vs factor by significance code (base)}
\usage{
plot_signif_base(df, x, y, signif, cols = 1:3)
}
\arguments{
\item{df}{Dataframe}

\item{x}{Unquoted x column}

\item{y}{Unquoted y column}

\item{signif}{Unquoted significance column}

\item{cols}{Colours [todo]}
}
\description{
Plot factor vs factor by significance code (base)
}
