% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signif}
\alias{plot_signif}
\title{Plot factor vs factor by significance code (ggplot2)}
\usage{
plot_signif(df, x, y, signif, pal = c(S = "red", NS = "green",
  `not-tested` = "grey50"))
}
\arguments{
\item{df}{Dataframe}

\item{x}{Unquoted x column}

\item{y}{Unquoted y column}

\item{signif}{Unquoted significance column}

\item{pal}{Colours [todo]}
}
\description{
Plot factor vs factor by significance code (ggplot2)
}
\examples{
plot_signif(ggplot2::mpg, manufacturer, class, drv,
pal = c("4" = "red", "f" = "blue", "r" = "orange"))

}
