% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signif_base}
\alias{plot_signif_base}
\title{Plot factor vs factor by significance code (base)}
\usage{
plot_signif_base(df, x, y, signif, pal = c(S = "red", NS = "green",
  `not-tested` = "grey50"), margins = c(3, 3))
}
\arguments{
\item{df}{Dataframe}

\item{x}{Unquoted x column}

\item{y}{Unquoted y column}

\item{signif}{Unquoted significance column}

\item{pal}{Colours for codes}

\item{margins}{vector of x and y margins for text}
}
\description{
Plot factor vs factor by significance code (base)
}
\examples{
plot_signif_base(ggplot2::mpg, manufacturer, class, drv,
pal = c("4" = "red", "f" = "blue", "r" = "orange"))

}
